+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 0.0.2 (2024--)
     Fixes:
 
-    * Fixed error message in ``Signal.chpoints`` setter.
+    * Error message in ``Signal.chpoints`` setter.
+    * Typing of ``Signal.y``.
 
 0.0.1 (2023-01-12)
     First release.